Added visualization of sold tickets to the ticket purchase window Added the ability to view statistics for a specific concert (number of tickets sold, box office receipts) Added the ability to terminate contracts and delete private accounts Added the ability for a singer to offer a contract to a manager Added messages tab

+++ Tickets_Consert_System.Data/Models/Client.cs
@@ -5,7 +5,7 @@ namespace Tickets_Consert_System
 {
     public sealed class Client : User
     {
-        public float Ballanse { get; set; }
+        public decimal Ballanse { get; set; }
 
         public Client() : base()
         {
